Extract repeat/synonym macro to own file
